Revert "no longer consider stat a read, which causes one test to fail (but will reduce redundant rebuilds)"
This reverts commit bb9560360f7e5f853c49953720b34daaae79eb92.

The reason is that I found that ignoring stat calls made fac rebuilds
far too fragile, particularly when using python, which caches a lot of
data in .pyc files.  Ignoring stats all too often meant we failed to
notice when we needed to rerun commands.  If python was only ever run
from fac, one might work around this, but it was too hard to maintain
that dependency information.