make fac ignore any files created outside the repository
This is a change in semantics that is moderately significant, although
it probably will not affect many repositories.  Previously, fac would
treat as an error creating or modifying any file outside the
repository directory.  Exceptions needed to be created using the `C`
directive.

This could in principle catch bugs such as a failure to remove a temp
file.  But in practice, it much more often caused me to document which
log files are created by which software, which is annoying and
error-prone.

With this change, fac simply ignores any files that are created
outside the repository.